// src/app/core/models/empresas.model.ts
export class Empresas {
    id: number;
    razaosocial: string;
    cpfoucnpj: string;
    naturezapessoa: string;
    cep: string;
    logradouro: string;
    numero: string;
    complemento: string;
    bairro: string;
    cidade: string;
    uf: string;
    nomecontato: string;
    telefone: string;
    whats: string;
    email: string;
    valor: number;
    datagravacao: Date;
    emailusuario: string;
    status: boolean;
  }
  
  